+++ packages/typir/src/kinds/function/function-validation-calls.ts
@@ -4,7 +4,7 @@
  * terms of the MIT License, which is available in the project root.
 ******************************************************************************/
 
-import { ReducedValidationProblem, ValidationProblem, ValidationProblemAcceptor, ValidationRuleLifecycle } from '../../services/validation.js';
+import { ValidationProblem, ValidationProblemAcceptor, ValidationRuleLifecycle } from '../../services/validation.js';
 import { TypirServices, TypirSpecifics } from '../../typir.js';
 import { RuleCollectorListener, RuleOptions } from '../../utils/rule-registration.js';
 import { TypirProblem } from '../../utils/utils-definitions.js';
@@ -102,7 +102,7 @@ export class FunctionCallArgumentsValidation<Specifics extends TypirSpecifics> i
             }
             // Since none of the function signatures match, report one validation issue (with sub-problems) for each function signature (and for each language key)
             if (resultOverloaded.length >= 1) {
-                accept(<ReducedValidationProblem<Specifics>>{
+                accept({
                     languageNode: languageNode,
                     severity: 'error',
                     message: `The given operands for the call of ${overloadedFunctions.overloadedFunctions.length >= 2 ? 'the overload ' : ''}'${overloadedName}' don't match.`,